Work done on showing a single recipe.
Added in a loading component. Tested it's feasability but not put into place yet.
Moved around the result components to their own folder so they aren't confused with the other search components.
Added in the component that will show a single recipe. Nothing has been done with this right now.