Adding loading, error and new recipe pages
Added in a loading page for when we fetch data.
Added in an error page if something was to explode.
Added in a new recipe page. This page works to a degree, in that a new recipe with name and description will be added. No validation is added, and the file upload does not work.